date filter to watchlist/list and /get controller

+++ src/backend_amirainvest_com/lib/backend_amirainvest_com/api/backend/watchlist/model.py
@@ -8,8 +8,8 @@ from common_amirainvest_com.models.creator import CreatorModel
 
 
 class WatchlistItemsReturnModel(BaseModel):
-    close_price: int
-    current_price: int
+    close_price: Decimal
+    current_price: Decimal
     ticker: str
     note: Optional[str]
     percent_change: Decimal
